got rid of a Borland warning

+++ Packages/Trapezoidal_decomposition/include/CGAL/Trapezoidal_decomposition_2.h
@@ -3026,8 +3026,8 @@ postcondition:
 		bool needs_update()
 		{
 			//to avoid g++ warning
-			signed long sz=number_of_curves();
-			if (rand()>RAND_MAX/(sz+1)) return false;
+			unsigned long sz=number_of_curves();
+			if ((unsigned) rand()> (RAND_MAX/(sz+1)) ) return false;
 			/*			 INTERNAL COMPILER ERROR overide
 			#ifndef __GNUC__
 			*/
